% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_coach.R
\name{bart_coach}
\alias{bart_coach}
\title{Get Coaching History and Notable Recruits}
\usage{
bart_coach(coach)
}
\arguments{
\item{coach}{Coach to return.}
}
\value{
Returns a list of tibbles: 'Coach History' and 'Recruiting History'
}
\description{
Returns coaching history with four factor and tournament finish (tibble one)
along with notable recruits (tibble two).
}
\details{
For a brief explanation of each factor and its computation, please visit
\href{https://kenpom.com/blog/four-factors/}{KenPom's blog}. `rec_rank` is a
scale of recruiting ranking, where 100 is a consensus #1 recruit and 0 is
unrecruited.
}
\examples{
bart_coach('Mike Krzyzewski')

}
