% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_factors.R
\name{bart_factors}
\alias{bart_factors}
\title{Get Four Factor Statistics}
\usage{
bart_factors(
  year = current_season(),
  result = NULL,
  type = NULL,
  start = NULL,
  end = NULL,
  location = NULL,
  last = NULL
)
}
\arguments{
\item{year}{Defaults to current season (YYYY).}

\item{result}{Filters by result ('W' or 'L')}

\item{type}{Filter by game type ('nc', 'conf', or 'post')}

\item{start}{Filters by starting date (YYYY-MM-DD)}

\item{end}{Filters by ending date (YYYY-MM-DD)}

\item{location}{Filters by game location ('H', 'A', or 'N')}

\item{last}{Filters by last x games played}
}
\value{

}
\description{
Returns four factor data and team records on a variety of splits, including
date range, quadrant level, opponent ranking, game location, and game type.
}
\details{
For a brief explanation of each factor and its computation, please visit
\href{https://kenpom.com/blog/four-factors/}{KenPom's blog}. Data can be
split on three variables: \describe{ \item{result}{Splits on game result: 'W'
or 'L'.} \item{type}{Splits on game type: 'nc' (non-conference), 'conf'
(conference), or 'post' (post-season tournaments).} \item{quad}{Splits by
quadrant level; 1-4 with 0 indicating 1-A games.} \item{top}{Splits by
opponent T-Rank position, adjusted for game location.}
\item{start/end}{Splits by date range (YYYYMMDD).}}
}
\examples{
\donttest{bart_factors(start='2022-01-13', type='conf')}

}
