% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{bart_game_box}
\alias{bart_game_box}
\title{Get Team Game Stats}
\usage{
bart_game_box(year = current_season())
}
\arguments{
\item{year}{Defaults to current season (YYYY).}
}
\description{
Returns box score data by team for all games.
}
\details{
The home team is coded as `team2`. Neutral site games may contain errors as
to whom is the home team.
}
\examples{
bart_game_box(year=2022)

}
