% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_players.R
\name{bart_players}
\alias{bart_players}
\title{Get All Players}
\usage{
bart_players(year = current_season())
}
\arguments{
\item{year}{Year to filter for; defaults to current season (YYYY)}
}
\value{
Returns a tibble with three columns:
\item{\code{player}}{character.}
\item{\code{team}}{character.}
\item{\code{id}}{double.}
\item{\code{year}}{double.}
}
\description{
Returns players and player IDs for a given season
}
\examples{
bart_players(year=2022)

}
