% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{bart_pregame}
\alias{bart_pregame}
\title{Get Pregame Probabilities}
\usage{
bart_pregame(year = current_season())
}
\arguments{
\item{year}{Defaults to current season (YYYY).}
}
\description{
Returns pregame win probabilities, expected scores, and thrill quotients for
all games.
}
\details{
Under the `type` column, games are classified by one of five indicators,
explained below: \describe{ \item{nc}{Non-conference games between two D-1
teams.} \item{conf}{In-conference games.} \item{conf_t}{Conference tournament
games.} \item{post}{Post-conference tournament games.} \item{nond1}{Games
involving one non-D1 team.} }
}
\examples{
\dontrun{bart_pregame(year=2022)}

}
