% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{bart_pro}
\alias{bart_pro}
\title{Get Pro Prospect Player Data}
\usage{
bart_pro(stat = "all", conf_only = FALSE, early = FALSE)
}
\arguments{
\item{stat}{Indicates statistical split (see details).}

\item{conf_only}{Logical. Filters data by conference-only play; defaults to
`FALSE`.}

\item{early}{Logical. Filters by early entrance status; defaults to FALSE
(all players).}
}
\value{
Returns a tibble with the number of columns dependent on the value
  supplied to the `stat` argument.
}
\description{
Returns detailed, season-long player statistics on a variety of splits for
all eligible or declared players.
}
\details{
Data is split on three statistical types, explained below: \describe{
\item{box}{Returns basic box score stats.} \item{shooting}{Returns
play-by-play shooting splits.} \item{adv}{Returns advanced metrics and
possession-adjusted box score statistics.}}
}
\examples{
bart_pro(stat='box')

}
