% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{bart_ratings}
\alias{bart_ratings}
\title{Get T-Rank Ratings}
\usage{
bart_ratings(year = current_season())
}
\arguments{
\item{year}{Defaults to current season (YYYY).}
}
\description{
Returns current T-Rank ratings and two forms of strength of schedule.
}
\details{
\itemize{\item Barthag is the estimation of a team's win probability against
the average Division 1 team on a neutral court. \item WAB (wins above bubble)
is the number of wins attained above or below the expected total from a
bubble team against an analogous schedule. \item `x_elite_sos` is the
percentage of games that an 'elite' team would project to lose against this
team's non-conference or overall schedule. \item `x_cur_sos` is the current
average Barthag rating of opponents. \item `x_fut_sos` is the projected
average Barthag rating of opponents.}
}
\examples{
bart_ratings(year=2022)

}
