% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{bart_team_history}
\alias{bart_team_history}
\title{Get Team History}
\usage{
bart_team_history(team)
}
\arguments{
\item{team}{Indicates team to return.}
}
\description{
Returns efficiency and four factor finish, conference and overall records,
and basic shooting percentages by year back to 2008.
}
\details{
`team` input must be an exact match: 'State' is abbreviated to 'St.' and full
names are used where applicable (e.g. 'North Carolina'). For complete list,
see `team` column of \code{\link{bart_ratings}}.
}
\examples{
bart_team_history(team='Charlotte')
}
