% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{bart_team_schedule}
\alias{bart_team_schedule}
\title{Get Team Schedule}
\usage{
bart_team_schedule(year = current_season(), team = NULL)
}
\arguments{
\item{year}{Defaults to current season (YYYY).}

\item{team}{Team to return.}
}
\value{
Returns a tibble of game-by-game four factors by team
}
\description{
Returns full team schedule and game-by-game four factor statistics by year
}
\details{
For a brief explanation of each factor and its computation, please visit
\href{https://kenpom.com/blog/four-factors/}{KenPom's blog}. `lead_diff` is
the average lead or deficit during a game for the team.
}
\examples{
bart_team_schedule(year=2022, team='Duke')

}
