% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{bart_team_shooting}
\alias{bart_team_shooting}
\title{Get Team Shooting Splits}
\usage{
bart_team_shooting(year = current_season())
}
\arguments{
\item{year}{Defaults to current season (YYYY).}
}
\description{
Returns team shooting totals and shares by shot location. Data returns back
to 2010.
}
\details{
`x_share` represents the percentage of made FGs that fall under that
category.
}
\examples{
bart_team_shooting(year=2019)
}
