% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_tournament_predicition.R
\name{bart_tournament_predicition}
\alias{bart_tournament_predicition}
\title{T-Rank Tournament Prediction}
\usage{
bart_tournament_predicition(
  teams = NULL,
  date = NULL,
  sims = NULL,
  seed = NULL
)
}
\arguments{
\item{teams}{List of teams arranged in scheduled order}

\item{date}{Date for tournament}

\item{sims}{Number of simulations to play}

\item{seed}{Seed to set for reproducible results}
}
\value{
Returns a tibble with four columns: \describe{
  \item{\code{team}}{character.} \item{\code{wins}}{integer.}
  \item{\code{finals}}{integer.} \item{\code{champ}}{integer.}}
}
\description{
Returns single-elimination tournament predictions using Barttorvik's
model
}
\details{
Given a list of teams arranged in scheduled order, function returns
tournament predictions for single-elimination tournaments
}
\examples{
\donttest{bart_tournament_predicition(teams=c('Duke', 'North Carolina',
'Kansas', 'Villanova'), date='20220402', sims=10, seed=1)}

}
