% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{bart_tourney_sheets}
\alias{bart_tourney_sheets}
\title{Get Committee Tournament Sheets}
\usage{
bart_tourney_sheets(year = current_season())
}
\arguments{
\item{year}{Defaults to current season (YYYY).}
}
\description{
Returns a tibble with similar quality and resume metrics used by the NCAA
seeding committee. Data runs back to the 2019 tournament.
}
\examples{
\dontrun{bart_tourney_sheets(year=2019)}

}
