% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Negate the base `%in%` operator}
\usage{
... \%!in\% NA
}
\description{
Reverses the base `%in%` operator and returns a boolean; TRUE if the
element(s) is not in the compared item
}
