% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_by_year.R
\name{teams_by_year}
\alias{teams_by_year}
\title{Teams and Conferences}
\usage{
teams_by_year(year = NULL, conf = NULL)
}
\arguments{
\item{year}{Year to filter}

\item{conf}{Conference to filter}
}
\value{
Returns a tibble with three columns:
\describe{
  \item{\code{team}}{character.} \item{\code{year}}{integer.}
  \item{\code{conf}}{character.}}
}
\description{
Returns teams and their conference affiliation by year
}
\details{
Function pulls teams and their men's basketball conference affiliation
by year back to 2008
}
\examples{
teams_by_year()

}
