% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_portal.R
\name{transfer_portal}
\alias{transfer_portal}
\title{Get Transfer Portal}
\usage{
transfer_portal(year = NULL, from = NULL, to = NULL)
}
\arguments{
\item{year}{Year to filter}

\item{from}{Team player is transferring from}

\item{to}{Team player is transferring to}
}
\value{
Returns a tibble with 31 columns: \describe{
  \item{\code{id}}{integer.} \item{\code{player}}{character.}
  \item{\code{from}}{character.} \item{\code{to}}{character.}
  \item{\code{exp}}{character.} \item{\code{year}}{integer.}
  \item{\code{imm_elig}}{character.} \item{\code{source}}{character.}
  \item{\code{from_d1}}{logical.} \item{\code{to_d1}}{logical}
  \item{\code{sit}}{logical.}}
}
\description{
Returns transfer portal decisions by year
}
\details{
Function pulls destination decisions by players in the transfer portal back
to 2012
}
\examples{
transfer_portal(to='Charlotte')

}
