% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{bart_archive}
\alias{bart_archive}
\title{Get T-Rank Archive Ratings}
\usage{
bart_archive(date)
}
\arguments{
\item{date}{Date to pull ratings (YYYYMMDD).}
}
\description{
Returns T-Rank ratings and efficiency metrics from the morning of the
specified day. Data goes back to 2014-15.
}
\details{
Barthag is the estimation of a team's win probability against the average
Division 1 team on a neutral court. WAB (wins above bubble) is the number of
wins attained above or below the expected total from a bubble team against an
analogous schedule.
}
\examples{
\dontrun{bart_archive('20220113')}
}
