% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_archive.R
\name{bart_archive}
\alias{bart_archive}
\title{Get T-Rank Archive Ratings}
\usage{
bart_archive(date = NULL, team = NULL, year = NULL)
}
\arguments{
\item{date}{Date to filter for (YYYYMMDD).}

\item{team}{Team to filter for.}

\item{year}{Year to filter for.}
}
\value{
Returns a tibble with 16 columns:
\describe{
  \item{\code{rk}}{double.}
  \item{\code{team}}{character.}
  \item{\code{conf}}{character.}
  \item{\code{rec}}{character.}
  \item{\code{barthag}}{double. The estimation of a team's win probability
  against the average Division 1 team on a neutral court.}
  \item{\code{adj_o}}{double.}
  \item{\code{adj_o_rk}}{double.}
  \item{\code{adj_d}}{double.}
  \item{\code{adj_d_rk}}{double.}
  \item{\code{adj_tempo}}{double.}
  \item{\code{adj_tempo_rk}}{double.}
  \item{\code{proj_rec}}{character.}
  \item{\code{proj_conf_rec}}{character.}
  \item{\code{wab}}{double. The number of wins above or below the expected
  total from a bubble team against the same schedule.}
  \item{\code{wab_rk}}{double.}
  \item{\code{date}}{double.}
}
}
\description{
Returns T-Rank ratings and efficiency metrics from the morning of the
specified day. Data goes back to 2014-15.
}
\examples{
\donttest{try(bart_archive(date='20220113'))}

}
