% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coach.R
\name{bart_coach_change}
\alias{bart_coach_change}
\title{Get Head Coach Changes}
\usage{
bart_coach_change(year = current_season())
}
\arguments{
\item{year}{Defaults to current season (YYYY).}
}
\description{
Returns head coaching changes at the Division 1 level by season.
}
\examples{
\dontrun{bart_coach_change(year=2015)}

}
