% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_coach_change.R
\name{bart_coach_change}
\alias{bart_coach_change}
\title{Get Head Coach Changes}
\usage{
bart_coach_change(year = current_season())
}
\arguments{
\item{year}{Defaults to current season (YYYY).}
}
\value{
Returns a tibble with four columns:
\describe{
  \item{\code{team}}{character.}
  \item{\code{conf}}{character.}
  \item{\code{old_coach}}{character.}
  \item{\code{new_coach}}{character.}
}
}
\description{
Returns head coaching changes at the Division 1 level by season.
}
\examples{
bart_coach_change(year=2015)

}
