% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{bart_conf_stats}
\alias{bart_conf_stats}
\title{Get Conference Team Statistics}
\usage{
bart_conf_stats(year = current_season(), conf = NULL)
}
\arguments{
\item{year}{Defaults to current season (YYYY).}

\item{conf}{Indicates conference (see details).}
}
\description{
Returns conference-only metrics, strengths of schedule, and bid probabilities
(auto and at-large) for all teams in a conference.
}
\details{
Accepted conference abbreviations for the `conf` argument are: \itemize{\item ‘A10’, ‘ACC’,
‘AE’, ‘ASun’, ‘Amer’, ‘B10’, ‘B12’, ‘BE’, ‘BSky’, ‘BSth’, ‘BW’, ‘CAA’,
‘CUSA’, ‘Horz’, ‘Ivy’, ‘MAAC’, ‘MAC’, ‘MEAC’, ‘MVC’, ‘MWC’, ‘NEC’, ‘OVC’,
‘P12’, ‘Pat’, ‘SB’, ‘SC’, ‘SEC’, ‘SWAC’, ‘Slnd’, ‘Sum’, ‘WAC’, ‘WCC’}

Other: \itemize{\item Barthag is the estimation of a team's win probability against
the average Division 1 team on a neutral court. \item `conf_cur_sos` is the
current average Barthag rating of conference opponents. \item `conf_fut_sos`
is the projected average Barthag rating of conference opponents.}
}
\examples{
\dontrun{bart_conf_stats(year=2022, conf='ACC')}

}
