% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_game_box.R
\name{bart_game_box}
\alias{bart_game_box}
\title{Get Team Game Stats}
\usage{
bart_game_box(
  year = current_season(),
  date = NULL,
  team = NULL,
  conf = NULL,
  game_id = NULL
)
}
\arguments{
\item{year}{Filters to year (YYYY)}

\item{date}{Filters to date (YYYY-MM-DD)}

\item{team}{Filters to team}

\item{conf}{Filters to conference}

\item{game_id}{Filters to game}
}
\value{
Returns a tibble of box score statistics
}
\description{
Returns box score data by team for all games.
}
\details{
The home team is coded as `team2`. Neutral site games may contain errors as
to whom is the home team.
}
\examples{
\donttest{try(bart_game_box(year=2022))}

}
