% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{bart_poy}
\alias{bart_poy}
\title{Get Player of the Year Ratings}
\usage{
bart_poy(year = current_season(), conf = "All", class = NULL, conf_only = F)
}
\arguments{
\item{year}{Defaults to current season (YYYY).}

\item{conf}{Filters results by conference; defaults to all (see details).}

\item{class}{Filters results by class ('fr', 'so', 'jr', 'sr'); defaults to
no filter.}

\item{conf_only}{Logical. Filters data by conference-only play; defaults to
`FALSE`.}
}
\description{
Returns Barttorvik Player of the Year ratings on a variety of splits.
}
\details{
Accepted conference abbreviations for the `conf` argument are: \itemize{\item ‘A10’, ‘ACC’,
‘AE’, ‘ASun’, ‘Amer’, ‘B10’, ‘B12’, ‘BE’, ‘BSky’, ‘BSth’, ‘BW’, ‘CAA’,
‘CUSA’, ‘Horz’, ‘Ivy’, ‘MAAC’, ‘MAC’, ‘MEAC’, ‘MVC’, ‘MWC’, ‘NEC’, ‘OVC’,
‘P12’, ‘Pat’, ‘SB’, ‘SC’, ‘SEC’, ‘SWAC’, ‘Slnd’, ‘Sum’, ‘WAC’, ‘WCC’ }
}
\examples{
\dontrun{bart_poy(year=2019, class='fr')}

}
