% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{bart_season_schedule}
\alias{bart_season_schedule}
\title{Get Full Season Schedule}
\usage{
bart_season_schedule(year = current_season())
}
\arguments{
\item{year}{Defaults to current season (YYYY).}
}
\description{
Returns full season schedule for all D-1 teams by year
}
\details{
Function pulls all games involving at least one D-1 team. Under the `type`
column, games are classified by one of four indicators, explained below:
\describe{ \item{nc}{Non-conference games between two D-1 teams.}
\item{conf}{In-conference games.} \item{post}{Post-conference tournament
games.} \item{nond1}{Games involving one non-D1 team.} }
}
\examples{
\dontrun{bart_season_schedule(year=2022)}

}
